refactor: 클래스명 및 위치 변경
