+++ android/app/src/main/java/com/daedan/festabook/FestaBookApp.kt
@@ -3,8 +3,8 @@ package com.daedan.festabook
 import android.app.Application
 import androidx.appcompat.app.AppCompatDelegate
 import com.daedan.festabook.logging.FirebaseAnalyticsTree
+import com.daedan.festabook.logging.FirebaseCrashlyticsTree
 import com.daedan.festabook.service.NotificationHelper
-import com.daedan.festabook.util.CrashlyticsTree
 import com.daedan.festabook.util.FestabookGlobalExceptionHandler
 import com.google.firebase.analytics.FirebaseAnalytics
 import com.naver.maps.map.NaverMapSdk
@@ -47,9 +47,9 @@ class FestaBookApp : Application() {
         if (BuildConfig.DEBUG) {
             plantDebugTimberTree()
         } else {
-          plantInfoTimberTree()
+            plantInfoTimberTree()
         }
-        Timber.plant(CrashlyticsTree())
+        Timber.plant(FirebaseCrashlyticsTree())
     }
 
     private fun plantDebugTimberTree() {
@@ -75,7 +75,13 @@ class FestaBookApp : Application() {
     }
 
     private fun setGlobalExceptionHandler() {
-        val defaultExceptionHandler: Thread.UncaughtExceptionHandler? = Thread.getDefaultUncaughtExceptionHandler()
-        Thread.setDefaultUncaughtExceptionHandler(FestabookGlobalExceptionHandler(this, defaultExceptionHandler))
+        val defaultExceptionHandler: Thread.UncaughtExceptionHandler? =
+            Thread.getDefaultUncaughtExceptionHandler()
+        Thread.setDefaultUncaughtExceptionHandler(
+            FestabookGlobalExceptionHandler(
+                this,
+                defaultExceptionHandler,
+            ),
+        )
     }
 }
